Split ProfileViewModel file into two files
